requirements.txt
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.6
cycler==0.10.0
dateparser==1.0.0
docutils==0.16
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.4.0
incremental==17.5.0
jeepney==0.6.0
joblib==1.0.1
keyring==22.0.1
kiwisolver==1.3.1
matplot==0.1.9
matplotlib==3.3.4
numpy==1.19.5
packaging==20.9
pandas==1.1.5
Pillow==8.1.0
pkginfo==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
PyHamcrest==2.0.2
pyloco==0.0.139
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-binance==0.7.9
python-dateutil==2.8.1
pytz==2021.1
readme-renderer==28.0
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
scikit-learn==0.24.1
scipy==1.5.4
SecretStorage==3.3.1
service-identity==18.1.0
SimpleWebSocketServer==0.1.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.57.0
twine==3.3.0
Twisted==20.3.0
txaio==21.2.1
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
ushlex==0.99.1
webencodings==0.5.1
websocket-client==0.57.0
zipp==3.4.0
zope.interface==5.2.0
